requirements_observability.txt
# Observability dependencies
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-instrumentation-psycopg2>=0.42b0
opentelemetry-instrumentation-redis>=0.42b0
opentelemetry-exporter-jaeger>=1.21.0
opentelemetry-exporter-jaeger-thrift>=1.21.0
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0
